refactored folders hierarchy
