+++ app/src/main/java/com/avelycure/photogallery/utils/FormatEditor.java
@@ -1,4 +1,4 @@
-package utils;
+package com.avelycure.photogallery.utils;
 
 import java.io.BufferedReader;
 import java.io.IOException;
