+++ app/src/main/java/com/avelycure/photogallery/utils/NetworkUtils.java
@@ -1,4 +1,4 @@
-package utils;
+package com.avelycure.photogallery.utils;
 
 import android.util.Log;
