+++ app/src/main/java/com/avelycure/photogallery/utils/PhotoGalleryDatabaseHelper.java
@@ -1,4 +1,4 @@
-package utils;
+package com.avelycure.photogallery.utils;
 
 import android.content.Context;
 import android.database.Cursor;
