+++ app/src/main/res/layout/acitvity_gallery.xml
@@ -5,7 +5,7 @@
     android:id="@+id/drawer"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PhotoGallery">
+    tools:context=".home.PhotoGallery">
 
             <LinearLayout
                 android:layout_width="match_parent"
