+++ app/src/main/res/layout/activity_albums.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AlbumsActivity">
+    tools:context=".albums.AlbumsActivity">
 <TextView
     android:layout_width="match_parent"
     android:layout_height="match_parent"
