+++ app/src/main/res/layout/activity_feedback.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".FeedbackActivity"
+    tools:context=".feedback.FeedbackActivity"
     android:background="@drawable/summer">
 
     <EditText
