+++ app/src/main/res/layout/activity_more.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MoreActivity">
+    tools:context=".more.MoreActivity">
     <TextView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
