+++ app/src/main/res/layout/activity_office.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".OfficeActivity">
+    tools:context=".ofiice.OfficeActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
